refactor: Remove task from project when deleting a task
This commit refactors the `deleteTaskService` in the `delete.service.ts` file to remove the task from the project when deleting a task. Instead of using the `filter` method, it now finds the index of the task in the `project.tasks` array and uses `splice` to remove it. This change improves the efficiency and accuracy of deleting tasks from projects.